requirements.txt
numpy>=1.23
torch>=2.1
pyyaml>=6.0
tensorboard>=2.14
pytest>=7.0
matplotlib>=3.7
